Move funding/reducer to fundingReducer
